% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackcess.R
\name{Database}
\alias{Database}
\title{Initialize a Database object instance for interacting with MS Access database files}
\usage{
Database(dsn, read_only = FALSE)
}
\arguments{
\item{dsn}{data source file path}

\item{read_only}{open connection as read-only? Default `FALSE`}
}
\value{
a com.healthmarketscience.jackcess.Database object
}
\description{
Initialize a Database object instance for interacting with MS Access database files
}
