% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackcess.R
\name{getColumnMetadata}
\alias{getColumnMetadata}
\title{Get Column Metadata}
\usage{
getColumnMetadata(db, table_names = NULL)
}
\arguments{
\item{db}{A \emph{com.healthmarketscience.jackcess.Database} object (e.g. created by \code{Database()})}

\item{table_names}{character. Database table name. Default \code{NULL} will return metadata for all tables in \code{db}.}
}
\value{
\emph{data.frame} containing database name (\code{database_name}; character), table name (\code{table_name}; character), column name (\code{column_name}; character), column data type (\code{datatype}; character), column length (\code{length}; integer), column length in units (\code{lengthinunits}; integer), column precision (\code{precision}; integer), column scale (\code{scale}; integer), column SQL type (\code{sqltype}; integer), column "is append only? (\code{appendonly}; logical), column "is autonumber?" (\code{autonumber}; logical), column "is calculated?" (\code{calculated}; logical), column "is compressed unicode?" (\code{compressedunicode}; logical), column "is hyperlink?" (\code{hyperlink}; logical), column "is variable length?" (\code{variablelength}; logical)
}
\description{
Get information about columns in specified table(s).
}
\seealso{
Database getTableNames
}
