% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackcess.R
\name{getTableNames}
\alias{getTableNames}
\title{Get Names of Tables in Database}
\usage{
getTableNames(db)
}
\arguments{
\item{db}{A \emph{com.healthmarketscience.jackcess.Database} object (e.g. created by \code{Database()})}
}
\value{
character
}
\description{
Get a character vector of all table names in specified database.
}
\seealso{
Database
}
