% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackcess.R
\name{importFile}
\alias{importFile}
\title{Import a delimited flat file into a new or existing MS Access Database table}
\usage{
importFile(
  dsn,
  db,
  name,
  delim = ",",
  quote = rJava::.jchar("\\""),
  filter = SimpleImportFilter()$INSTANCE,
  useExistingTable = FALSE,
  header = TRUE
)
}
\arguments{
\item{dsn}{Input file path}

\item{db}{a Database object}

\item{name}{Table name}

\item{delim}{Delimiter (default (\verb{,}))}

\item{quote}{Quote character (default \verb{"})}

\item{filter}{Optional custom ImportFilter object}

\item{useExistingTable}{logical. Use existing tables? Default: \code{FALSE}}

\item{header}{logical. Input file contains headers? Default: \code{TRUE}}
}
\description{
Import a delimited flat file into a new or existing MS Access Database table
}
