% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjackcess-package.R
\docType{package}
\name{rjackcess-package}
\alias{rjackcess}
\alias{rjackcess-package}
\title{rjackcess: Interface to Jackcess, a Java library for reading from and writing to MS Access databases}
\description{
rjackcess uses rJava to provide a basic interface to the Jackcess \url{https://jackcess.sourceforge.io/} Java API for MS Access databases. More information on the API is available in the Javadocs (\url{https://javadoc.io/doc/com.healthmarketscience.jackcess/jackcess/latest/index.html}).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://humus.rocks/rjackcess/}
}

}
\keyword{internal}
