% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackcess.R
\name{SimpleExportFilter}
\alias{SimpleExportFilter}
\title{SimpleExportFilter}
\usage{
SimpleExportFilter()
}
\value{
A \emph{com.healthmarketscience.jackcess.util.SimpleImportFilter} object
}
\description{
SimpleExportFilter
}
