% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackcess.R
\name{SimpleImportFilter}
\alias{SimpleImportFilter}
\title{SimpleImportFilter}
\usage{
SimpleImportFilter()
}
\value{
a com.healthmarketscience.jackcess.util.SimpleImportFilter object
}
\description{
SimpleImportFilter
}
