% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackcess.R
\name{Database}
\alias{Database}
\title{Initialize a Jackcess \code{Database} object instance for interacting with MS Access database files}
\usage{
Database(
  dsn,
  read_only = FALSE,
  file_format = c("V2019", "V2016", "V2010", "V2007", "V2003", "V2000", "V1997",
    "MSISAM", "GENERIC_JET4"),
  overwrite = FALSE
)
}
\arguments{
\item{dsn}{\emph{character}. MS Access database source name (file path).}

\item{read_only}{\emph{logical} Open connection as read-only? Default \code{FALSE}}

\item{file_format}{\emph{character}. One of the following: \code{"GENERIC_JET4"}, \code{"MSISAM"}, \code{"V1997"}, \code{"V2000"}, \code{"V2003"}, \code{"V2007"}, \code{"V2010"}, \code{"V2016"}, \code{"V2019"}. See details.}

\item{overwrite}{\emph{logical}. overwrite existing file? Default \code{FALSE}}
}
\value{
A \emph{com.healthmarketscience.jackcess.Database} object
}
\description{
Initialize a Jackcess \code{Database} object instance for interacting with MS Access database files
}
\details{
\itemize{
\item \code{"GENERIC_JET4"} - A database which was most likely created programmatically
\item \code{"MSISAM"} - A database which was created by MS Money
\item \code{"V1997"} - A database which was created by MS Access 97
\item \code{"V2000"} - A database which was created by MS Access 2000
\item \code{"V2003"} - A database which was created by MS Access 2002/2003
\item \code{"V2007"} - A database which was created by MS Access 2007
\item \code{"V2010"} - A database which was created by MS Access 2010+
\item \code{"V2016"} - A database which was created by MS Access 2016+
\item \code{"V2019"} - A database which was created by MS Access 2019+ (Office 365)
}
}
