% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackcess.R
\name{exportAll}
\alias{exportAll}
\title{Export contents of a MS Access Database to Flat Files}
\usage{
exportAll(db, dsn)
}
\arguments{
\item{db}{a Database object}

\item{dsn}{directory path for output files}
}
\description{
Export contents of a MS Access Database to Flat Files
}
